Fix QML module version in generated qmldir
We still used the decade-old 2.0 version to generate qmldir and qmltypes.
This has caused issues issues with ecm_find_qmlmodule()

Also, get rid of the remaining references to kirigami.2 directory with
version suffix. We should no longer be using it.

Note that I had to manually run qmlplugindump, because generated target
does not pass `-noinstantiate` option (without which qmlplugindump
includes a lot of unrelated stuff from QtQuick types), and does not use
a full path to binary (and without altered PATH on my system it can only
find Qt 5 version).